+++ Utilities/vtkexodus2/exppem.c
@@ -155,7 +155,7 @@ int ex_put_partial_elem_map (int exoid,
   }
 
   /* Check input parameters for a valid range of numbers */
-  if (ent_start <= 0 || ent_start > num_elem) {
+  if (ent_start <= 0 || (size_t)ent_start > num_elem) {
     exerrval = EX_FATAL;
     sprintf(errmsg,
 	    "Error: start count is invalid in file id %d",
@@ -171,7 +171,7 @@ int ex_put_partial_elem_map (int exoid,
     ex_err("ex_put_partial_elem_map",errmsg,exerrval);
     return (EX_FATAL);
   }
-  if (ent_start+ent_count-1 > num_elem) {
+  if ((size_t)(ent_start+ent_count-1) > num_elem) {
     exerrval = EX_FATAL;
     sprintf(errmsg,
 	    "Error: start+count-1 is larger than element count in file id %d",
